refactor: create src folder
