+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "type": "module",
   "scripts": {
-    "start": "node index.mjs",
-    "start:user": "node index.mjs --username=Eugene"
+    "start": "node src/index.mjs",
+    "start:user": "node src/index.mjs --username=Eugene"
   },
   "keywords": [],
   "author": "",
