+++ src/utils/directory.mjs
@@ -1,0 +1,8 @@
+const getCurrentPath = () => {
+  return process.cwd();
+}
+
+const changePath = (newPath) => {
+  process.chdir(newPath);
+}
+export { getCurrentPath, changePath };
