--- utils/directory.mjs
@@ -1,8 +1,0 @@
-const getCurrentPath = () => {
-  return process.cwd();
-}
-
-const changePath = (newPath) => {
-  process.chdir(newPath);
-}
-export { getCurrentPath, changePath };
